annotatedappapi-master/requirements.txt
click==7.1.2
dataclasses==0.6
fastapi==0.61.1
h11==0.11.0
pydantic==1.7.2
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.12.2
